# TDB/src/main/java/alaaesam/tdb/model/vto/CurrencyVTO.java
package alaaesam.tdb.model.vto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Builder
@AllArgsConstructor
@NoArgsConstructor
@Data
public class CurrencyVTO {
    private String name;
    private String symbol;
}